requirements.txt
Flask==2.1.1
openai~=0.27.2
apig-wsgi==2.15.0
python-dotenv~=1.0.0
boto3==1.20.16

